ADD DATA ANALYTICS MODULE
Replaces 'diagnostics' with 'analytics' throughout level 1 field-based experiments, renames and updates analytics modules, and moves the static charge profile plotting to a new analytics module. Updates parameter files and launcher logic to use the new analytics toggles and module. Removes the old displacement profile plot and adds a new charge profile plot in the appropriate directory.
