+++ openwave/xperiments/4_level0_granule_based/_xparameters/spacetime_vibration.py
@@ -57,7 +57,7 @@ XPARAMETERS = {
         "COLOR_PALETTE": 0,  # Color palette list: default (99), granule-type (0), redshift (1), ironbow (2), blueprint (3), viridis (4)
     },
     "analytics": {
-        "ANALYTICS": False,  # Toggle analytics (speed & wavelength measurements)
+        "ANALYTICS": False,  # Toggle data analytics (speed & wavelength measurements)
         "EXPORT_VIDEO": False,  # Toggle frame image export to video directory
         "VIDEO_FRAMES": 24,  # Target frame number to stop recording and finalize video export
     },
